Analytics_Dash/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
executing==0.8.3
gunicorn==20.1.0
idna==3.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.23.1
oauthlib==3.2.0
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
